fix(test): the two gates that were hiding behind conformance_diff_error
8e76c9413 fixed the first non-compiling target in the abi suite. Re-running the
check with --keep-going instead of letting it stop at the first failure showed
two more behind it. That is the abort chain's real cost: without --keep-going I
would have found these one per build cycle, each looking like a fresh breakage.

conformance_diff_ns_rdata_types
    The fl arm passed `c_int::from(ty)` to an export declared `u16`. Fixed by
    passing `ty` unconverted -- but the mismatch is not a test typo. fl declares
    `class: u16, ty: u16, ttl: u32` where libresolv's prototype is
    `ns_class, ns_type, u_long`: enums, so `int`, and 64-bit. The gate's own type
    alias is the corroboration, not my reading of the header -- its author wrote
    `c_int, c_int, c_ulong` and the arm runs green against live libresolv.

    Benign today: these are register-passed integer arguments, so a narrower
    callee reads the low half of the register the caller filled, and RR class and
    type are 16-bit on the wire anyway. It is still a divergent C signature.
    Filed as bd-n70esc rather than fixed here, because widening a production
    export forces a decision about out-of-range values at the boundary that
    belongs to the resolv owner, not to a test fix.

conformance_diff_wcsftime_l
    Three errors. The `tm` parameter is now `*const c_void` so fl's export and
    glibc's unify with one alias -- fl declares it that way, and the two are
    ABI-identical since a pointer is a pointer.

    `libc::tzset` does not exist, and rustc's suggested
    `frankenlibc_abi::time_abi::tzset` is the SECOND instance today of a
    suggestion that would quietly hollow a gate: it updates FL's zone state,
    while the arm that reads the process zone for `%Z` is GLIBC's. fl is UTC-only
    and needs no tzset, so glibc's is both necessary and sufficient, resolved
    through the shared dlsym oracle with fl's own export as the guard.

    In a differential suite the helpful suggestion is the dangerous one: it makes
    the error disappear by pointing the oracle at the implementation under test.

VERIFIED, one invocation, clean overlay naming both files:
  rch exec --base HEAD --clean-overlay --overlay-path <each>
    -- cargo check -p frankenlibc-abi --all-targets --keep-going
  rc=0, ZERO compile errors.

So the abi suite compiles end to end again: three broken targets total, and
nothing was hiding behind these two. `cargo test -p frankenlibc-abi` can
enumerate its targets instead of dying during compilation, which is the whole
point -- an aborted suite reports nothing at all rather than reporting failure.

NOT YET RUN. The follow-up `cargo test` for these two gates has been queued
behind a contended fleet (`queue_timeout`, then no admission) and has not
produced a result. Committing on the compile evidence deliberately: these targets
do not build on main today, so any outcome here is strictly better, and if either
gate turns out RED at runtime that is a visible failure belonging to bd-aykfv1
rather than a silent abort. The runtime result will be recorded there.

Refs: bd-aykfv1, bd-n70esc, bd-jksmc2, bd-9kuafg, bd-v0388t

+++ crates/frankenlibc-abi/tests/conformance_diff_ns_rdata_types.rs
@@ -95,12 +95,20 @@ fn render_both(ty: u16) -> (c_int, c_int) {
             host_buf.as_mut_ptr(),
             host_buf.len(),
         );
+        // `ty` unconverted, because fl's export declares `class: u16, ty: u16,
+        // ttl: u32` while libresolv's prototype (SprintrrfFn above) is
+        // `ns_class, ns_type, u_long` -- enums, so `int`, and `unsigned long`.
+        // Both arms therefore receive the SAME VALUE through different declared
+        // widths. On x86-64 SysV that is benign for values this gate passes: the
+        // argument register holds the value and a narrower callee reads its low
+        // half. It is still a divergent C signature; filed rather than changed
+        // here, because widening a production export is not a test fix.
         let fl_rc = frankenlibc_abi::resolv_abi::ns_sprintrrf(
             msg.as_ptr(),
             msg.len(),
             name.as_ptr(),
             1,
-            c_int::from(ty),
+            ty,
             0,
             rdata.as_ptr(),
             rdata.len(),
